Split success and failure cases into separate values
